refactor: move classes to archiver package
As it's related to archiving I think it fits here better.
